+++ Packaging/homebrew/Casks/updatebar-app.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 cask "updatebar-app" do
-  version "0.3.1"
+  version "0.3.2"
   sha256 "32a2ae9cdf0d37e0cb8c909d9a4537e959d3a61c94e75e9e90a61fe5f7d5da67"
 
   url "https://github.com/sonim1/UpdateBar/releases/download/v#{version}/UpdateBar-#{version}-macos-arm64.app.tar.gz"